+++ db/migrations/20220513184841_create_nvdicts_table.sql
@@ -1,7 +1,7 @@
 -- +micrate Up
-CREATE TABLE IF NOT EXISTS nvdicts (
+CREATE TABLE IF NOT EXISTS nvdicts(
   id serial PRIMARY KEY,
-  nvinfo_id bigint NOT NULL,
+  nvinfo_id int NOT NULL,
   --
   dname text NOT NULL DEFAULT '',
   d_lbl text NOT NULL DEFAULT '',
@@ -22,19 +22,19 @@ CREATE TABLE IF NOT EXISTS nvdicts (
   updated_at timestamptz NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
 
-CREATE UNIQUE INDEX nvdict_unique_idx ON nvdicts (dname);
+CREATE UNIQUE INDEX nvdict_unique_idx ON nvdicts(dname);
 
-CREATE INDEX nvdict_nvinfo_idx ON nvdicts (nvinfo_id);
+CREATE INDEX nvdict_nvinfo_idx ON nvdicts(nvinfo_id);
 
-CREATE INDEX nvdict_parent_idx ON nvdicts (parent_id);
+CREATE INDEX nvdict_parent_idx ON nvdicts(parent_id);
 
-CREATE INDEX nvdict_unames_idx ON nvdicts USING GIN (unames);
+CREATE INDEX nvdict_unames_idx ON nvdicts USING GIN(unames);
 
-CREATE INDEX nvdict_themes_idx ON nvdicts USING GIN (themes);
+CREATE INDEX nvdict_themes_idx ON nvdicts USING GIN(themes);
 
-CREATE INDEX nvdict_utime_idx ON nvdicts (utime);
+CREATE INDEX nvdict_utime_idx ON nvdicts(utime);
 
-CREATE INDEX nvdict_dsize_idx ON nvdicts (dsize);
+CREATE INDEX nvdict_dsize_idx ON nvdicts(dsize);
 
 -- +micrate Down
 DROP TABLE IF EXISTS nvdicts;
